backend/requirements.txt
flask==2.0.1
werkzeug==2.0.1
pymongo
pyjwt
openai>=1.0.0
python-dotenv
flask-cors
sentry-sdk
blinker
SpeechRecognition
PyAudio